+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+- New plugin config structure. Object map replaced with an array of objects, having at least a `uri` key, an optional `config` key and other keys such as `translations`. See [IPluginConfigMeta](./src/IPluginConfigMeta.ts).
+- Support translation overrides per plugin configuration (`translations` key)
+
 # v1.0.0-beta.8
 
 - Allow optional adapters dependencies in modules to not throw an error if the adapter is undefined.